windows: recent win10 support Unix Domain
Allow selection of Unix Domain Sockets on windows since it is supported
for the last couple of years on windows 10

https://devblogs.microsoft.com/commandline/af_unix-comes-to-windows/

... if only they could add a full set of posix pieces to go with it
(and abstract namespace UDS which doesn't work apparently) so that
the parts dealing with uid / gid don't have to be disabled.

+++ lib/roles/http/client/client-handshake.c
@@ -681,9 +681,9 @@ ads_known:
 	if (wsi->unix_skt) {
 		psa = (const struct sockaddr *)&sau;
 		if (sau.sun_path[0])
-			n = sizeof(uint16_t) + strlen(sau.sun_path);
+			n = (int)(sizeof(uint16_t) + strlen(sau.sun_path));
 		else
-			n = sizeof(uint16_t) + strlen(&sau.sun_path[1]) + 1;
+			n = (int)(sizeof(uint16_t) + strlen(&sau.sun_path[1]) + 1);
 	} else
 #endif
